backend/delivery_service/requirements.txt
django==4.2.27
djangorestframework==3.16.1
psycopg2-binary==2.9.10
pika==1.3.2
requests==2.32.3
pytest==8.3.4
pytest-django==4.9.0